edagames-server/requirements.txt
PyJWT==2.0.1
requests==2.25.1
fastapi==0.63.0
starlette==0.13.6
uvicorn==0.13.4
h11==0.12.0
click==7.1.2
pydantic==1.8.2
typing-extensions==3.7.4.3
websockets==8.1
edagames-grpc==1.1.0
redis==3.5.3
